Extract `simpleHash` function from oh-repeater to `$oh.utils`
